59 find out why discard card fields are sometimes bugged

+++ Assets/Scripts/CardScripts/ShopBehaviour.cs
@@ -121,7 +121,7 @@ public class ShopBehaviour : NetworkBehaviour
         deckManager.buyAnyCard = false;
     }
 
-    private ErrorMsg BuyCard(GameObject card, int coins)
+    private ErrorMsg BuyCard(GameObject card, float coins)
     {
         CardBehaviour cardBehaviour = card.GetComponent<CardBehaviour>();
         if (coins == cardBehaviour.price)
